feat(avatar): refactor code
