+++ core/components/atoms/selectionAvatarGroup/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './SelectionAvatarGroup';
+export * from './SelectionAvatarGroup';
