+++ core/components/atoms/selectionAvatarGroup/selectionAvatarPopover/SelectionAvatarInput.tsx
