+++ core/components/atoms/selectionAvatarGroup/selectionAvatarPopover/SelectionAvatarList.tsx
