+++ core/components/atoms/selectionAvatarGroup/selectionAvatarPopover/SelectionAvatarOption.tsx
