+++ core/components/atoms/selectionAvatarGroup/selectionAvatarPopover/index.tsx
@@ -1,0 +1,1 @@
+export * from './SelectionAvatarPopover';
