+++ core/components/atoms/selectionAvatarGroup/selectionAvatars/SelectionAvatar.tsx
