+++ core/components/atoms/selectionAvatarGroup/selectionAvatars/index.tsx
@@ -1,0 +1,1 @@
+export * from './SelectionAvatars';
